Påskinlämning

+++ app.rb
@@ -15,7 +15,7 @@ end
 get('/') do
   db = SQLite3::Database.new("db/database.db")
   db.results_as_hash = true
-  custom_timers_result = db.execute("SELECT * FROM custom_timers")
+  custom_timers_result = db.execute("SELECT * FROM custom_timers WHERE user_id = ?", session[:user_id])
   default_timers_result = db.execute("SELECT * FROM default_timers")
   #puts "hello eeadedqwqdwq #{default_timers_result}"
 
@@ -29,16 +29,57 @@ get('/') do
 end 
 
 get('/default_timers/new/') do
+  if session[:permission] != "admin"
+    set_error("Permission denied")
+    redirect("/")
+  end
+
   slim(:"default_timers/new")
 end 
-get('/default_timers/edit/') do
-  slim(:"default_timers/edit")
+post('/default_timers/new/') do
+  name = params[:name]
+  hour = 0
+  minute = params[:minute]
+
+  db = SQLite3::Database.new("db/database.db")
+  db.execute("INSERT INTO default_timers (name, hour, minute) VALUES (?, ?, ?)", [name, hour, minute])
+  redirect('/')
+end
+get('/default_timers/:time/edit/') do
+  if session[:permission] != "admin"
+    set_error("Permission denied")
+    redirect("/")
+  end
+
+  hour = params[:time].split("X")[0]
+  minute = params[:time].split("X")[1]
+
+  db = SQLite3::Database.new("db/database.db")
+  db.results_as_hash = true
+  result = db.execute("SELECT * FROM default_timers WHERE hour = ? AND minute = ?", [hour, minute]).first
+  puts "Results: #{result}"
+  slim(:"/default_timers/edit", locals:{timer:result})
 end 
+
+post('/default_timers/:time/update/') do
+  puts "Hello 2: #{params[:time]}"
+  currentHour = params[:time].split("X")[0]
+  currentMinute = params[:time].split("X")[1]
+  name = params[:name]
+  hour = params[:hour]
+  minute = params[:minute]
+
+  puts " Hello: #{[currentHour, currentMinute, name, hour, minute]}"
+  db = SQLite3::Database.new("db/database.db")
+  db.results_as_hash = true
+  db.execute("UPDATE default_timers SET hour = ?, minute = ?, name = ? WHERE hour = ? AND minute = ?", [hour, minute, name, currentHour, currentMinute])
+  redirect('/')
+end
 
 get('/timers/') do
   db = SQLite3::Database.new("db/database.db")
   db.results_as_hash = true
-  custom_timers_result = db.execute("SELECT * FROM custom_timers")
+  custom_timers_result = db.execute("SELECT * FROM custom_timers WHERE user_id = ?", [session[:user_id]])
 
   slim(:"timers/index", locals:{custom_timers:custom_timers_result})
 end
@@ -55,24 +96,37 @@ post('/timers/new/') do
 
   db = SQLite3::Database.new("db/database.db")
   db.execute("DELETE FROM custom_timers WHERE hour = ? AND minute = ? AND user_id = ?", [hour, minute, user_id])
-  db.execute("INSERT INTO custom_timers (name, hour, minute, user_id) VALUES (?, ?, ?)", [name, hour, minute, user_id])
+  db.execute("INSERT INTO custom_timers (name, hour, minute, user_id) VALUES (?, ?, ?, ?)", [name, hour, minute, user_id])
   redirect('/timers/')
 end 
 
 post('/timers/:time/delete/') do
   hour = params[:time].split("X")[0]
   minute = params[:time].split("X")[1]
+  user_id = session[:user_id]
+
   db = SQLite3::Database.new("db/database.db")
-  db.execute("DELETE FROM custom_timers WHERE hour = ? AND minute = ?", [hour, minute])
+  db.execute("DELETE FROM custom_timers WHERE hour = ? AND minute = ? AND user_id = ?", [hour, minute, user_id])
   redirect('/timers/')
+end 
+
+post('/default_timers/:time/delete/') do
+  hour = params[:time].split("X")[0]
+  minute = params[:time].split("X")[1]
+
+  db = SQLite3::Database.new("db/database.db")
+  db.execute("DELETE FROM default_timers WHERE hour = ? AND minute = ?", [hour, minute])
+  redirect('/')
 end 
 
 get('/timers/:time/edit/') do
   hour = params[:time].split("X")[0]
   minute = params[:time].split("X")[1]
+  user_id = session[:user_id]
+
   db = SQLite3::Database.new("db/database.db")
   db.results_as_hash = true
-  result = db.execute("SELECT * FROM custom_timers WHERE hour = ? AND minute = ?", [hour, minute]).first
+  result = db.execute("SELECT * FROM custom_timers WHERE hour = ? AND minute = ? AND user_id = ?", [hour, minute, user_id]).first
   puts "Results: #{result}"
   slim(:"/timers/edit", locals:{timer:result})
 end 
@@ -83,10 +137,11 @@ post('/timers/:time/update/') do
   name = params[:name]
   hour = params[:hour]
   minute = params[:minute]
+  id = session[:user_id]
   puts " Hello: #{[currentHour, currentMinute, name, hour, minute]}"
   db = SQLite3::Database.new("db/database.db")
   db.results_as_hash = true
-  db.execute("UPDATE custom_timers SET hour = ?, minute = ?, name = ? WHERE hour = ? AND minute = ?", [hour, minute, name, currentHour, currentMinute])
+  db.execute("UPDATE custom_timers SET hour = ?, minute = ?, name = ? WHERE hour = ? AND minute = ? AND user_id = ?", [hour, minute, name, currentHour, currentMinute, id])
   redirect('/timers/')
 end
